Move user_cache_dir fixture to conftest.py
