+++ tests/test_cache.py
@@ -1,21 +1,9 @@
 """Tests for the cache module."""
 from pathlib import Path
 
-import pytest
-from _pytest.monkeypatch import MonkeyPatch
-
 from .utils import commit
 from cutty import cache
 from cutty import git
-
-
-@pytest.fixture
-def user_cache_dir(monkeypatch: MonkeyPatch, tmp_path: Path) -> Path:
-    """Replace the application cache directory by a temporary directory."""
-    path = tmp_path / ".cache" / cache.appname
-    path.mkdir(parents=True, exist_ok=True)
-    monkeypatch.setattr("appdirs.user_cache_dir", lambda *args, **kwargs: path)
-    return path
 
 
 def test_repository_clones(user_cache_dir: Path, repository: git.Repository) -> None:
